lmkd: count the number of times LMKD wakes up
LMKD wake up when the memory pressure was high enough either in either the psi or the vmpressure mechanism. The  memory reclaim ability is based on CPUs' capcaity of a chip and it is different from each others. This patch can count the number of lmkd wake up when meet the memory pressure threshold instead of heavy-loading logging. To show the count, we can just reinit the lmkd, and it will print in the android logcat and reset the count to zero.

Test: Run APP rotation

Bug:365748420

Signed-off-by: JohnHsu <john.hsu@mediatek.com>
Change-Id: I3980d2a90a910c64449b4ad2b005e4d0437097e8